Fixing Entities, Libraries, and Program

+++ src/Ca.Backend.Test.Domain/Entities/BillingLineEntity.cs
@@ -5,9 +5,8 @@ public class BillingLineEntity : BaseEntity
     public BillingEntity Billing { get; set; }
     public Guid ProductId { get; set; }
     public ProductEntity Product { get; set; }
-    public decimal Quantity { get; set; }
+    public int Quantity { get; set; }
     public decimal UnitPrice { get; set; }
     public decimal Subtotal { get; set; }
-    
 }
